+++ skillfarm/helpers/discord.py
@@ -12,14 +12,12 @@ from allianceauth.authentication.models import User
 from allianceauth.notifications import notify
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.logging import LoggerAddTag
-
 # AA Skillfarm
 from skillfarm import __title__
 from skillfarm.constants import DISCORD_EMBED_COLOR_MAP
+from skillfarm.providers import AppLogger
 
-logger = LoggerAddTag(get_extension_logger(__name__), __title__)
+logger = AppLogger(my_logger=get_extension_logger(__name__), prefix=__title__)
 
 
 def allianceauth_discordbot_installed() -> bool: